refactor: fix remain error when changing money datatype from double to BigDecimal, complete io data tool and refactor model to match with io feature

+++ src/main/java/vn/id/nonglam/kltn/kltn/dto/response/order/OrderResponse.java
@@ -1,9 +1,10 @@
 package vn.id.nonglam.kltn.kltn.dto.response.order;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
-public record OrderResponse(boolean status, UUID orderId, double deposit) {
+public record OrderResponse(boolean status, UUID orderId, BigDecimal deposit) {
     public record RoomDetailSnapShotResponse(UUID id, String code, boolean valid) {}
 
     public record RoomDetailValidResponse(UUID roomTypeId, String name,
